pyproject.toml
[tool.poetry]
name = "leasecheck-canada"
version = "0.1.0"
description = ""
authors = ["John Smith <john@example.com>"]

packages = [
   { include = "leasecheck" }
]


[tool.poetry.dependencies]
python = ">=3.11,<3.12"
python-dotenv = ">=1.1.1"
requests = ">=2.32.4"
numpy = ">=1.26.4"
pandas = ">=2.3.0"
scipy = ">=1.13.1"
torch = ">=2.7.1"
jupyterlab = ">=4.4.3"
ipywidgets = ">=8.1.7"
matplotlib = ">=3.10.3"
scikit-learn = ">=1.7.0"
chromadb = ">=1.0.12"
jupyter-dash = ">=0.4.2"
sentencepiece = ">=0.2.0"
pyarrow = ">=20.0.0"
beautifulsoup4 = ">=4.13.4"
plotly = ">=6.2.0"
bitsandbytes = ">=0.46.0"
transformers = ">=4.53.0"
sentence-transformers = ">=4.1.0"
datasets = ">=3.6.0"
accelerate = ">=1.8.1"
openai = ">=1.92.2"
anthropic = ">=0.55.0"
google-generativeai = ">=0.8.5"
gradio = ">=5.34.2"
gensim = ">=4.3.3"
modal = ">=1.0.4"
ollama = ">=0.5.1"
psutil = ">=7.0.0"
setuptools = ">=80.9.0"
speedtest-cli = ">=2.1.3"
langchain = ">=0.3.26"
langchain-core = ">=0.3.66"
langchain-text-splitters = ">=0.3.8"
langchain-openai = ">=0.3.26"
langchain-chroma = ">=0.2.4"
langchain-community = ">=0.3.26"
feedparser = ">=6.0.11"
twilio = ">=9.6.3"
pydub = ">=0.25.1"
pdfplumber = "^0.11.7"
streamlit = "^1.46.1"
ruff = "^0.12.1"
black = "^25.1.0"
faiss-cpu = ">=1.8.0"     

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
summarize = "scripts.run_cli:main"